ft: rename the go pkg.
