+++ service/feed.go
@@ -1,9 +1,10 @@
 package service
 
 import (
-	"douyin/dao"
 	"net/http"
 	"time"
+
+	"github.com/whyy1/douyin/dao"
 
 	"github.com/gin-gonic/gin"
 )
